Moving MeasureDisplay into containers
